pyproject.toml
[project]
name = "deepresearch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.13",
    "anthropic>=0.54.0",
    "asyncio>=3.4.3",
    "duckduckgo-search>=8.0.4",
    "gradio[mcp]>=5.34.0",
    "llama-index>=0.12.42",
    "llama-index-tools-mcp>=0.2.5",
    "llama-index-utils-workflow>=0.3.3",
    "nebius>=0.2.34",
    "openai>=1.86.0",
    "pytest>=8.4.0",
    "pytest-asyncio>=1.0.0",
    "python-dotenv>=1.1.0",
    "serpapi>=0.1.5",
    "tavily-python>=0.7.6",
    "types-aiofiles>=24.1.0.20250606",
    "types-pyyaml>=6.0.12.20250516",
    "types-requests>=2.32.4.20250611",
]
